start using spaces for assets

+++ app/Http/Controllers/ImageController.php
@@ -2,12 +2,11 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
 use App\Image;
-use Validator;
-use Illuminate\Support\Facades\Storage;
-use Illuminate\Http\File;
 use App\Item;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
+use Validator;
 
 class ImageController extends Controller
 {
@@ -22,26 +21,30 @@ class ImageController extends Controller
             return ['code' => '1', 'msg' => $validator->errors()->first()];
         }
 
-        if ($request->item_id == 'null') {
+        if ('null' == $request->item_id) {
             return ['code' => '1', 'msg' => 'Item ID is empty'];
         }
 
-
         foreach ($request->file('files') as $image) {
-            $saveImage = basename(Storage::putFile('public/item_images', $image));
+            // $saveImage = basename(Storage::putFile('public/item_images', $image));
+            $saveImage = basename(Storage::disk('spaces')->putFile('item_images', $image));
             Image::create(['name' => $saveImage, 'item_id' => $request->item_id]);
         }
-        // return $a;
         $items = Item::all();
+
         return ['code' => '0', 'msg' => 'ok', 'result' => $items];
     }
 
     public function destroy($id)
     {
         $image = Image::find($id);
-        Storage::delete($image->name);
-        $image->delete();
+        // Storage::delete($image->name);
+        // $image->delete();
+        if (Storage::disk('spaces')->delete('item_images/'.$image->name)) {
+            $image->delete();
+        }
         $items = Item::all();
+
         return ['code' => '0', 'msg' => 'ok', 'result' => $items];
     }
 }
